Fix size_t/int casting problem on Win64 build
TBR=jkummerow@chromium.org

Review URL: https://codereview.chromium.org/11364121

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12880 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/deoptimizer.cc
@@ -106,7 +106,7 @@ static const int kDeoptTableMaxEpilogueCodeSize = 2 * KB;
 
 
 size_t Deoptimizer::GetMaxDeoptTableSize() {
-  size_t entries_size =
+  int entries_size =
       Deoptimizer::kMaxNumberOfEntries * Deoptimizer::table_entry_size_;
   int page_count = ((kDeoptTableMaxEpilogueCodeSize + entries_size - 1) /
                     OS::CommitPageSize()) + 1;
